modeltraining/requirements.txt
joblib==1.5.1
pandas==2.2.1
fsspec==2025.5.1
gcsfs==2025.5.1
google-cloud-storage==3.1.0
scikit-learn==1.6.1